[Buildbot] Create buildbot_utils.py for buildbot specific functions (#282)
(cherry picked from commit ec84e5cb3c0e11ed8bc8290e5ab1c6a8f9701773)
